+++ Source/NSURLSessionConfiguration+TNLAdditions.m
@@ -13,51 +13,45 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@implementation NSURLSessionConfiguration (TNLAdditions)
+static struct {
+    BOOL URLSessionCanReceiveResponseViaDelegate:1;
+    BOOL URLSessionCanUseTaskTransactionMetrics:1;
+    BOOL URLSessionSupportsDecodingBrotliContentEncoding:1;
+    BOOL URLSessionCanUseWaitsForConnectivity:1;
+} sFlags;
 
-+ (BOOL)tnl_URLSessionCanReceiveResponseViaDelegate
+static void _EnsureFlags(void);
+static void _EnsureFlags()
 {
-    static BOOL sBugExists;
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
+
+        memset(&sFlags, 0, sizeof(sFlags)); // clear the flags before we set them
+
+        /// URLSessionCanReceiveResponseViaDelegate
 
         if (tnl_available_ios_9) {
             // ok
+            sFlags.URLSessionCanReceiveResponseViaDelegate = YES;
         } else {
-            // iOS 8 only has this bug
-            sBugExists = YES;
+            // iOS 8 only has this bug (TNL does not support iOS 7 or below)
         }
 
-    });
-
-    return !sBugExists;
-}
-
-+ (BOOL)tnl_URLSessionCanUseTaskTransactionMetrics
-{
-    static BOOL sTaskMetricsAvailable = NO;
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
+        /// URLSessionCanUseTaskTransactionMetrics
 
         if (tnl_available_ios_10) {
             // task metrics added as an API in iOS 10
 
             if (tnl_available_ios_11) {
-                // Crashers of iOS 10 continue into iOS 11 betas
-                // Cannot differentiate iOS 11.0.0 and iOS 11 betas
-                // So...
-                //   On iOS 11.0.1, consider fixed
-                //   On iOS 11.0.0, consider bug present
-                //   On non-iOS targets, just presume non-beta and consider fixed
-#if TARGET_OS_IOS
-                if (@available(iOS 11.0.1, *)) {
-                    // definitely fixed on iOS 11.0.1
-                    sTaskMetricsAvailable = YES;
-                }
-#else
-                // non-iOS, consider fixed
-                sTaskMetricsAvailable = YES;
-#endif
+
+                // The crashers of iOS 10 continue into iOS 11 betas.
+                // The crash was fixed in iOS 11.0.0 GM.
+                // Since we cannot differentiate GM vs earlier beta, there can be crashing...
+                // ...but those betas are old enough now to not need special consideration.
+
+                // Consider fixed on iOS 11+
+                sFlags.URLSessionCanUseTaskTransactionMetrics = YES;
+
             } else {
                 // task metrics exist but have crashes on iOS 10.X
                 // iOS 10.0.X and iOS 10.1.X have a crashing bug that crashes 1 million times a day
@@ -65,16 +59,7 @@ NS_ASSUME_NONNULL_BEGIN
             }
         }
 
-    });
-
-    return sTaskMetricsAvailable;
-}
-
-+ (BOOL)tnl_URLSessionSupportsDecodingBrotliContentEncoding
-{
-    static BOOL sBrotliSupported = NO;
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
+        /// URLSessionSupportsDecodingBrotliContentEncoding
 
         // Brotli support requires 2 things:
         // - Running OS version of iOS 11 (or equivalent platform version) or greater
@@ -83,13 +68,50 @@ NS_ASSUME_NONNULL_BEGIN
 
 #if TARGET_SDK_SUPPORTS_BROTLI
         if (tnl_available_ios_11) {
-            sBrotliSupported = YES;
+            sFlags.URLSessionSupportsDecodingBrotliContentEncoding = YES;
         }
 #endif
 
-    });
+        /// URLSessionCanUseWaitsForConnectivity
 
-    return sBrotliSupported;
+        if (tnl_available_ios_11) {
+
+            // added iOS 11
+            sFlags.URLSessionCanUseWaitsForConnectivity = YES;
+
+            if (tnl_available_ios_13) {
+
+                // regressed iOS 13
+                sFlags.URLSessionCanUseWaitsForConnectivity = NO;
+            }
+        }
+    });
+}
+
+@implementation NSURLSessionConfiguration (TNLAdditions)
+
++ (BOOL)tnl_URLSessionCanReceiveResponseViaDelegate
+{
+    _EnsureFlags();
+    return sFlags.URLSessionCanReceiveResponseViaDelegate;
+}
+
++ (BOOL)tnl_URLSessionCanUseTaskTransactionMetrics
+{
+    _EnsureFlags();
+    return sFlags.URLSessionCanUseTaskTransactionMetrics;
+}
+
++ (BOOL)tnl_URLSessionSupportsDecodingBrotliContentEncoding
+{
+    _EnsureFlags();
+    return sFlags.URLSessionSupportsDecodingBrotliContentEncoding;
+}
+
++ (BOOL)tnl_URLSessionCanUseWaitsForConnectivity
+{
+    _EnsureFlags();
+    return sFlags.URLSessionCanUseWaitsForConnectivity;
 }
 
 - (void)tnl_insertProtocolClasses:(nullable NSArray<Class> *)additionalClasses